Support Auto-Agree
The fact that server software expects you to have stdin bound to a
terminal says a lot about the engineering quality of the project.
Since there's no way to do what every other acme client does and agree
to the EULA in a file, we version the service file instead and inject
some magic flags.